// src/api/generated/model/handlersRankMatchHistoryResult.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * BFH Auth Example API
 * Backend API for BFH
 * OpenAPI spec version: 1.0
 */

export interface HandlersRankMatchHistoryResult {
  added_damage?: number;
  at?: string;
  attacker?: number;
  battle_id?: number;
  defender?: number;
  last_rate?: number;
  new_rate?: number;
  taken_damage?: number;
  win?: boolean;
  win_uft?: number;
}